Anchor to window, not ui
Update default window_config function to anchor the window to the top
right of the current window, not the top right of the nvim ui.

Realized it doesn't make sense to pass a ui object to the window_config
function, so that parameter has been removed. If this object is required
to figure out the desired window configuration it can be obtained with
the following command: vim.api.nvim_list_uis()[1]